+++ packages/core/src/odata-v4/request-builder/get-by-key-request-builder.spec.ts
@@ -49,16 +49,14 @@ describe('GetByKeyRequestBuilder', () => {
 
       mockGetRequest(
         {
-          path: "A_TestEntityWithEnumKey(KeyPropertyEnum1='Member1',KeyPropertyEnum2='Member2',KeyPropertyEnum3='Member1')",
+          path: "A_TestEntityWithEnumKey(KeyPropertyEnum1='Member1')",
           responseBody: entityData
         },
         TestEntityWithEnumKey
       );
 
       const actual = await new GetByKeyRequestBuilder(TestEntityWithEnumKey, {
-        KeyPropertyEnum1: expected.keyPropertyEnum1,
-        KeyPropertyEnum2: expected.keyPropertyEnum2,
-        KeyPropertyEnum3: expected.keyPropertyEnum3
+        KeyPropertyEnum1: expected.keyPropertyEnum1
       }).execute(defaultDestination);
       expect(actual).toEqual(expected);
     });