+++ packages/generator/src/enum-type/enum.ts
@@ -1,16 +1,16 @@
 import { EnumDeclarationStructure, StructureKind } from 'ts-morph';
 import { VdmEnumType } from '../vdm-types';
-import { enumDocs } from '../typedoc';
+import { addLeadingNewline, enumDocs } from '../typedoc';
 
 export function enumTypeClass(enumType: VdmEnumType): EnumDeclarationStructure {
   return {
     kind: StructureKind.Enum,
     name: enumType.typeName,
     isExported: true,
-    members: Object.keys(enumType.members).map(key => ({
-      name: key,
-      value: key,
-      docs: ['test', 'test2']
+    members: enumType.members.map(member => ({
+      name: member.name,
+      value: member.name,
+      docs: [addLeadingNewline(`original value: ${member.originalValue}`)]
     })),
     docs: [enumDocs(enumType)]
   };