+++ test-packages/test-services/v4/multiple-schemas-service/TestComplexType2.ts
@@ -3,7 +3,17 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import { ComplexTypeField, ConstructorOrField, EdmTypeField, EntityV4, FieldBuilder, FieldOptions, FieldType, PropertyMetadata, deserializeComplexTypeV4 } from '@sap-cloud-sdk/core';
+import {
+  ComplexTypeField,
+  ConstructorOrField,
+  EdmTypeField,
+  EntityV4,
+  FieldBuilder,
+  FieldOptions,
+  FieldType,
+  PropertyMetadata,
+  deserializeComplexTypeV4
+} from '@sap-cloud-sdk/core';
 
 /**
  * TestComplexType2
@@ -26,13 +36,18 @@ export function createTestComplexType2(json: any): TestComplexType2 {
  * TestComplexType2Field
  * @typeparam EntityT - Type of the entity the complex type field belongs to.
  */
-export class TestComplexType2Field<EntityT extends EntityV4, NullableT extends boolean = false, SelectableT extends boolean = false> extends ComplexTypeField<EntityT, TestComplexType2, NullableT, SelectableT> {
+export class TestComplexType2Field<
+  EntityT extends EntityV4,
+  NullableT extends boolean = false,
+  SelectableT extends boolean = false
+> extends ComplexTypeField<EntityT, TestComplexType2, NullableT, SelectableT> {
   private _fieldBuilder: FieldBuilder<this> = new FieldBuilder(this);
   /**
    * Representation of the [[TestComplexType2.stringProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
    */
-  stringProperty: EdmTypeField<EntityT, 'Edm.String', false, false> = this._fieldBuilder.buildEdmTypeField('StringProperty', 'Edm.String', false);
+  stringProperty: EdmTypeField<EntityT, 'Edm.String', false, false> =
+    this._fieldBuilder.buildEdmTypeField('StringProperty', 'Edm.String', false);
 
   /**
    * Creates an instance of TestComplexType2Field.
@@ -40,7 +55,11 @@ export class TestComplexType2Field<EntityT extends EntityV4, NullableT extends b
    * @param fieldName - Actual name of the field as used in the OData request.
    * @param fieldOf - Either the parent entity constructor of the parent complex type this field belongs to.
    */
-  constructor(fieldName: string, fieldOf: ConstructorOrField<EntityT>, fieldOptions?: FieldOptions<NullableT, SelectableT>) {
+  constructor(
+    fieldName: string,
+    fieldOf: ConstructorOrField<EntityT>,
+    fieldOptions?: FieldOptions<NullableT, SelectableT>
+  ) {
     super(fieldName, fieldOf, TestComplexType2, fieldOptions);
   }
 }
@@ -49,12 +68,14 @@ export namespace TestComplexType2 {
   /**
    * Metadata information on all properties of the `TestComplexType2` complex type.
    */
-  export const _propertyMetadata: PropertyMetadata<TestComplexType2>[] = [{
-    originalName: 'StringProperty',
-    name: 'stringProperty',
-    type: 'Edm.String',
-    isCollection: false
-  }];
+  export const _propertyMetadata: PropertyMetadata<TestComplexType2>[] = [
+    {
+      originalName: 'StringProperty',
+      name: 'stringProperty',
+      type: 'Edm.String',
+      isCollection: false
+    }
+  ];
 
   /**
    * @deprecated Since v1.25.0. Use `deserializeComplexTypeV2` or `deserializeComplexTypeV4` of the `@sap-cloud-sdk/core` package instead.