+++ test-packages/test-services/v4/multiple-schemas-service/TestEnumType1.js
@@ -1,25 +1,24 @@
-"use strict";
+'use strict';
 /*
  * Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved.
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-Object.defineProperty(exports, "__esModule", { value: true });
+Object.defineProperty(exports, '__esModule', { value: true });
 exports.TestEnumType1 = void 0;
 /**
- * Enum type: [[A_TestEnumType1]] that reflects all the members in the metadata.
- * The underlying type of this enum is Edm.Int32
- * The value of the enum entries are:
- * Member1: 1
- * Member2: 2
+ * This enum represents the enum type [[A_TestEnumType1]]".
+ * The members represent values of EDM type Edm.Int32.
  */
 var TestEnumType1;
 (function (TestEnumType1) {
-    /** test */
-    /** test2 */
-    TestEnumType1["Member1"] = "Member1";
-    /** test */
-    /** test2 */
-    TestEnumType1["Member2"] = "Member2";
-})(TestEnumType1 = exports.TestEnumType1 || (exports.TestEnumType1 = {}));
+  /**
+   * original value: 1
+   */
+  TestEnumType1['Member1'] = 'Member1';
+  /**
+   * original value: 2
+   */
+  TestEnumType1['Member2'] = 'Member2';
+})((TestEnumType1 = exports.TestEnumType1 || (exports.TestEnumType1 = {})));
 //# sourceMappingURL=TestEnumType1.js.map