+++ test-packages/test-services/v4/test-service/TestEnumType.js
@@ -1,25 +1,24 @@
-"use strict";
+'use strict';
 /*
  * Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved.
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-Object.defineProperty(exports, "__esModule", { value: true });
+Object.defineProperty(exports, '__esModule', { value: true });
 exports.TestEnumType = void 0;
 /**
- * Enum type: [[A_TestEnumType]] that reflects all the members in the metadata.
- * The underlying type of this enum is Edm.Int32
- * The value of the enum entries are:
- * Member1: 0
- * Member2: 1
+ * This enum represents the enum type [[A_TestEnumType]]".
+ * The members represent values of EDM type Edm.Int32.
  */
 var TestEnumType;
 (function (TestEnumType) {
-    /** test */
-    /** test2 */
-    TestEnumType["Member1"] = "Member1";
-    /** test */
-    /** test2 */
-    TestEnumType["Member2"] = "Member2";
-})(TestEnumType = exports.TestEnumType || (exports.TestEnumType = {}));
+  /**
+   * original value: 0
+   */
+  TestEnumType['Member1'] = 'Member1';
+  /**
+   * original value: 1
+   */
+  TestEnumType['Member2'] = 'Member2';
+})((TestEnumType = exports.TestEnumType || (exports.TestEnumType = {})));
 //# sourceMappingURL=TestEnumType.js.map