+++ test-packages/test-services/v4/test-service/TestEnumTypeWithOneMember.js
@@ -1,21 +1,24 @@
-"use strict";
+'use strict';
 /*
  * Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved.
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-Object.defineProperty(exports, "__esModule", { value: true });
+Object.defineProperty(exports, '__esModule', { value: true });
 exports.TestEnumTypeWithOneMember = void 0;
 /**
- * Enum type: [[A_TestEnumType_WithOneMember]] that reflects all the members in the metadata.
- * The underlying type of this enum is Edm.Int32
- * The value of the enum entries are:
- * Member1: 0
+ * This enum represents the enum type [[A_TestEnumType_WithOneMember]]".
+ * The members represent values of EDM type Edm.Int32.
  */
 var TestEnumTypeWithOneMember;
 (function (TestEnumTypeWithOneMember) {
-    /** test */
-    /** test2 */
-    TestEnumTypeWithOneMember["Member1"] = "Member1";
-})(TestEnumTypeWithOneMember = exports.TestEnumTypeWithOneMember || (exports.TestEnumTypeWithOneMember = {}));
+  /**
+   * original value: 0
+   */
+  TestEnumTypeWithOneMember['Member1'] = 'Member1';
+})(
+  (TestEnumTypeWithOneMember =
+    exports.TestEnumTypeWithOneMember ||
+    (exports.TestEnumTypeWithOneMember = {}))
+);
 //# sourceMappingURL=TestEnumTypeWithOneMember.js.map